CopyrightStatementsProcessor: Make `process()` deterministic
Due to the grouping by normalized owner, parts with actually different
owners could map to the same key. When merging parts, it was actually
depending on the order of the parts which of the potentially different
owners was used. Fix that by first sorting before grouping. This gives
shorter owners precedence, which seems reasonable as these are likely
more reabable.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>